pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "specsmith"
version = "0.3.0a1"
description = "Applied Epistemic Engineering toolkit — forge epistemically-governed scaffolds, stress-test belief systems, and run AEE pipelines."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "BitConcepts"},
]
keywords = [
    "agentic", "scaffold", "governance", "agents-md", "cli",
    "epistemic-engineering", "belief-artifacts", "stress-testing",
    "certainty", "trace-vault", "aee", "knowledge-engineering",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.0",
    "types-pyyaml>=6.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]
# LLM provider extras for specsmith run (agentic client)
anthopic = ["anthropic>=0.56"]
openai = ["openai>=1.0"]
gemini = ["google-generativeai>=0.8"]
# Install all optional LLM providers
agent = ["anthropic>=0.56", "openai>=1.0"]
# Convenience bundle: everything
all = [
    "anthropic>=0.56",
    "openai>=1.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
specsmith = "specsmith.cli:main"

[project.urls]
Homepage = "https://github.com/BitConcepts/specsmith"
Documentation = "https://specsmith.readthedocs.io"
Repository = "https://github.com/BitConcepts/specsmith"
Changelog = "https://github.com/BitConcepts/specsmith/blob/main/CHANGELOG.md"
Issues = "https://github.com/BitConcepts/specsmith/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
specsmith = ["templates/**/*.j2", "agent/profiles/*.md"]
epistemic = ["py.typed"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
